Add consumer module
Separates the log consumption side from the serving side, since those
can now share state via the CRDT. The consumer reads both data and log
topics. It has the ability to pause/resume data topics to have the right
number of records prepared for consumption.

It should gain the ability to pause all data topics in order to catch up
on the log - it doesn't have that yet!